# api/openapi.yaml
openapi: 3.0.3
info:
  title: PulseControlERP API
  description: Multi-tenant ERP system for garment manufacturing and supply chain management
  version: 1.0.0
  contact:
    name: PulseControlERP Support
servers:
  - url: https://api.pulsecontroleRP.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
          nullable: true
        roleCode:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        isActive:
          type: boolean
    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
    MasterOrder:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        internalOrderNo:
          type: string
        buyerOrderNo:
          type: string
        buyerName:
          type: string
        styleNo:
          type: string
        orderQty:
          type: number
        pcdDate:
          type: string
          format: date
        bpcdDate:
          type: string
          format: date
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
...
  /companies:
    get:
      tags: [System Admin]
      summary: List all companies (System Admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Companies list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
                  total:
                    type: integer

    post:
      tags: [System Admin]
      summary: Create new company (System Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: Company created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
  /companies/{id}:
    get:
      tags: [System Admin]
      summary: Get company details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'

  # ==================== Users ====================
  /users:
    get:
      tags: [Users]
      summary: List users (filtered by company for non-admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'

    post:
      tags: [Users]
      summary: Create new user (Company Admin or System Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, fullName, roleId]
              properties:
                companyId:
                  type: integer
                  format: int64
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
                fullName:
                  type: string
                roleId:
                  type: integer
                  format: int64
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== Master Orders ====================
  /orders:
    get:
      tags: [Merchandising]
      summary: List master orders
      parameters:
        - name: buyerOrderNo
          in: query
          schema:
            type: string
        - name: internalOrderNo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MasterOrder'

    post:
      tags: [Merchandising]
      summary: Create master order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [internalOrderNo, buyerOrderNo, buyerName, orderQty]
              properties:
                internalOrderNo:
                  type: string
                buyerOrderNo:
                  type: string
                buyerName:
                  type: string
                styleNo:
                  type: string
                orderQty:
                  type: number
                pcdDate:
                  type: string
                  format: date
                bpcdDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterOrder'
  /orders/{id}:
    get:
      tags: [Merchandising]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterOrder'

  # ==================== Sub-PO ====================
...